blob: remove clusters checksums when data change
When a blob with a child snapshot is deleted or when a snapshot is
inflated or decoupled from its parent, snapshot's data could change.
So we remove the snapshot clusters checksums, if present, to prevent the
reading of obsolete values.

Longhorn 10037

Signed-off-by: Damiano Cipriani <damiano.cipriani@suse.com>
(cherry picked from commit ff4f219b578412fde5041c672d9f85459be76d65)